# swagger.json
{"openapi":"3.1.0","info":{"title":"HappyEduHub API","version":"1.0.0","description":"API của HappyEduHub"},"tags":[{"name":"Auth","description":"Các API liên quan đến xác thực."},{"name":"Tags","description":"Các API liên quan đến tag."},{"name":"Users","description":"Các API liên quan đến người dùng."}],"components":{"schemas":{"BulkCreateTagDto":{"type":"array","items":{"$ref":"#/components/schemas/CreateTagDto"},"minItems":1},"BulkDeleteTagDto":{"type":"array","items":{"type":"string","format":"uuid"},"minItems":1},"BulkUpdateTagDto":{"type":"array","items":{"$ref":"#/components/schemas/UpdateTagDto"},"minItems":1},"CreateTagDto":{"type":"object","properties":{"name":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"type":{"type":"string","enum":["SUBJECT","GRADE"]},"fragment":{"type":"string","minLength":1}},"required":["name","description","type","fragment"]},"Tag":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"type":{"type":"string","enum":["SUBJECT","GRADE"]},"fragment":{"type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"},"version":{"type":"number"}},"required":["id","name","description","type","fragment","created_at","updated_at","version"]},"TagArray":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"UpdateTagDto":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string","minLength":1},"description":{"type":"string","minLength":1},"type":{"type":"string","enum":["SUBJECT","GRADE"]},"fragment":{"type":"string","minLength":1}},"required":["id"]},"User":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","ACADEMIC_AFFAIR","STUDENT","TEACHER"]},"gender":{"type":"string","enum":["MALE","FEMALE"]},"avatar":{"type":["string","null"]},"tel":{"type":["string","null"]},"dob":{"type":["string","null"]},"code":{"type":["string","null"]},"address":{"type":["string","null"]},"created_at":{"type":"string"},"updated_at":{"type":"string"},"version":{"type":"number"}},"required":["id","email","role","gender","created_at","updated_at","version"]}},"parameters":{},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register a new user","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":1},"gender":{"type":"string","enum":["MALE","FEMALE"]},"avatar":{"type":"string"},"tel":{"type":"string"},"dob":{"type":"string"},"code":{"type":"string"},"address":{"type":"string"}},"required":["email","password","gender"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login to the system","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6}},"required":["email","password"]}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/tags":{"get":{"tags":["Tags"],"summary":"Lấy tất cả các tag.","responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagArray"}}}}}},"post":{"tags":["Tags"],"summary":"Tạo tag mới.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkCreateTagDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagArray"}}}}}},"put":{"tags":["Tags"],"summary":"Cập nhật tag.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkUpdateTagDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagArray"}}}}}},"delete":{"tags":["Tags"],"summary":"Xóa tag.","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkDeleteTagDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]}}}}}}},"/users/profile":{"get":{"tags":["Users"],"summary":"Get user profile","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}}},"webhooks":{}}